@@ -28,14 +28,14 @@
   },
   "dependencies": {
     "@4c/cli-core": "^2.3.0",
-    "@babel/cli": "^7.12.13",
-    "@babel/core": "^7.12.13",
+    "@babel/cli": "^7.15.4",
+    "@babel/core": "^7.15.5",
     "@manypkg/get-packages": "^1.1.1",
-    "cpy": "^8.1.1",
-    "execa": "^5.0.0",
-    "fs-extra": "^9.1.0",
+    "cpy": "^8.1.2",
+    "execa": "^5.1.1",
+    "fs-extra": "^10.0.0",
     "listr": "^0.14.3",
-    "typescript": "^4.1.3"
+    "typescript": "^4.4.3"
   },
   "publishConfig": {
     "access": "public"
