@@ -1,9 +1,9 @@
 import path from 'path';
 import { promisify } from 'util';
 
-import ConsoleUtilities from '@4c/cli-core/ConsoleUtilities';
-import GitUtilities from '@4c/cli-core/GitUtilities';
-import PromptUtilities from '@4c/cli-core/PromptUtilities';
+import * as ConsoleUtilities from '@4c/cli-core/ConsoleUtilities';
+import * as GitUtilities from '@4c/cli-core/GitUtilities';
+import * as PromptUtilities from '@4c/cli-core/PromptUtilities';
 import { createAltPublishDir } from '@4c/file-butler';
 import exitHook from 'async-exit-hook';
 import chalk from 'chalk';
@@ -11,7 +11,7 @@ import execa from 'execa';
 import fs from 'fs-extra';
 import hasYarn from 'has-yarn';
 import Listr from 'listr';
-import readPkgUp from 'read-pkg-up';
+import { readPackageUpAsync } from 'read-pkg-up';
 import rimraf from 'rimraf';
 import { from } from 'rxjs';
 import { catchError } from 'rxjs/operators';
@@ -219,7 +219,7 @@ export const builder = (_) =>
 
 const handlerImpl = async (argv) => {
   const cwd = process.cwd();
-  const { path: pkgPath, packageJson } = await readPkgUp({
+  const { path: pkgPath, packageJson } = await readPackageUpAsync({
     cwd,
     normalize: false,
   });
